BrickTest

Exercises

Exercise 5

list3 :: (Ord n, Show n) => SimpleApp (List n String) n

Similar to list2, but now we use Brick. The selected item should be highlighted, the arrow keys should change the selection. The Enter key should end the application.

Brick already supplies us with a List Widget, so have a look at functions like renderList and handleListEvent.

Keep in mind that everything will be renderd in default style unless you define an attrMap!

Bonus: center the list horizontally and vertically on the screen.

runSimpleBrickApp list3 list3TestState >>= print . listSelectedElement

list3TestState :: List String String

Exercise 6

dialogExample :: SimpleApp (Dialog Bool) ()

A Dialog asking the user to answer »Yes« or »No«.

Brick supplies us with a Dialog Widget, have a look at renderDialog and handleDialogEvent.

The dialog should be rendered in reverse video, with the active button in red.

runSimpleBrickApp dialogExample yesNoDialog >>= print . dialogSelection

yesNoDialog :: Dialog Bool

Exercise 7

renderPongPaddle :: (Int, Int) -> Widget n

Define your own Widget.

Create a widget that renders a Pong paddle. The function takes as parameters the offset of the top edge of the paddle, and the height of the paddle. The widget should be unbounded (Greedy) in height, and have a width of exactly 1. For the render function, use emptyResult and add an image using the Vty render functions in Graphics.Vty.Image.

pongPaddleExample :: SimpleApp (Int, Int) ()

Integrate the pong paddle widget into an app.

Arrow keys should move the paddle up and down, Esc should end the application.

runSimpleBrickApp pongPaddleExample (10, 10)

Utilities

data SimpleApp s n

A simplified version of Brick's App type.

SimpleApp s e n is an application with state s and internal names of type n.

Constructors

SimpleApp

Fields

simpleAppDraw :: s -> [Widget n]

The drawing function. Widgets are stacked, the first item in the list appears on top.

simpleAppDraw :: String -> [Widget n]
simpleAppDraw = s -> [str s]

simpleAppHandleEvent :: s -> BrickEvent n () -> EventM n (Next s)

Handle a Brick event. The EventM type allows access to Brick's internal state (e.g. getVtyHandle :: EventM n Vty). To determine the next action, use one of:

continue :: s -> EventM n (Next s) to resume with an updated application state

halt :: s -> EventM n (Next s) to end the app with a given final state

suspendAndResume :: IO s -> EventM n (Next s) to suspend the application, run the given IO action, and resume with the new application state.

simpleAppHandleEvent :: Int -> BrickEvent n () -> EventM n (Next s)
simpleAppHandleEvent = n e -> case e of
    VtyEvent (EvKey KUp [])   -> continue (n + 1)
    VtyEvent (EvKey KDown []) -> continue (n - 1)
    VtyEvent (EvKey KEsc [])  -> halt n
    _otherwise                -> continue n

simpleAppAttrMap :: s -> AttrMap

Define a state-dependent AttrMap. An AttrMap is constructed using attrMap :: Attr -> [(AttrName, Attr)] -> AttrMap, where the first argument is the default attribute (usually defAttr), and the second argument are overrides for special attributes. For AttrNames, see documentation of the individual Widgets.

runSimpleBrickApp :: Ord n => SimpleApp s n -> s -> IO s

Runs a SimpleApp full-screen and prints the final state after the application is ended.

let app :: SimpleApp () ()
    app = SimpleApp
        { simpleAppDraw = _ -> [str "Hello World"]
        , simpleAppHandleEvent = _ _ -> halt ()
        , simpleAppAttrMap = _ -> attrMap defAttr [] }
in runSimpleBrickApp app ()

Library Functions

List Widget

data List n e :: * -> * -> *

List state. Lists have an element type e that is the data stored by the list. Lists handle the following events by default:

Up/down arrow keys: move cursor of selected item

Page up / page down keys: move cursor of selected item by one page at a time (based on the number of items shown)

Home/end keys: move cursor of selected item to beginning or end of list

list

Arguments

:: n
The list name (must be unique)

-> Vector e
The initial list contents

-> Int
The list item height in rows (all list item widgets must be this high)

-> List n e

Construct a list in terms of an element type e.

renderList

Arguments

:: (Ord n, Show n)

=> (Bool -> e -> Widget n)
Rendering function, True for the selected element

-> Bool
Whether the list has focus

-> List n e
The List to be rendered

-> Widget n
rendered widget

Turn a list state value into a widget given an item drawing function.

handleListEvent :: Ord n => Event -> List n e -> EventM n (List n e)

Dialog Widget

data Dialog a :: * -> *

Dialogs present a window with a title (optional), a body, and buttons (optional). Dialog buttons are labeled with strings and map to values of type a, which you choose.

Dialogs handle the following events by default with handleDialogEvent:

Tab or Right Arrow: select the next button

Shift-tab or Left Arrow: select the previous button

dialog

Arguments

:: Maybe String
The dialog title

-> Maybe (Int, [(String, a)])
The currently-selected button index (starting at zero) and the button labels and values to use

-> Int
The maximum width of the dialog

-> Dialog a

Create a dialog.

renderDialog :: Dialog a -> Widget n -> Widget n

Render a dialog with the specified body widget. This renders the dialog as a layer, which makes this suitable as a top-level layer in your rendering function to be rendered on top of the rest of your interface.

handleDialogEvent :: Event -> Dialog a -> EventM n (Dialog a)

Basic Widgets

str :: String -> Widget n

Build a widget from a String. Breaks newlines up and space-pads short lines out to the length of the longest line.

center :: Widget n -> Widget n

Center a widget both vertically and horizontally. Consumes all available vertical and horizontal space.

hCenter :: Widget n -> Widget n

Center the specified widget horizontally. Consumes all available horizontal space.

padAll :: Int -> Widget n -> Widget n

Pad a widget on all sides. Defers to the padded widget for growth policy.

AttrMaps

attrMap

Arguments

:: Attr
The map's default attribute to be returned when a name lookup fails, and the attribute that will be merged with successful lookups.

-> [(AttrName, Attr)]
The map's initial contents.

-> AttrMap

Create an attribute map.

listAttr :: AttrName

The top-level attribute used for the entire list.

listSelectedAttr :: AttrName

The attribute used only for the currently-selected list item when the list does not have focus. Extends listAttr.

dialogAttr :: AttrName

The default attribute of the dialog

buttonAttr :: AttrName

The default attribute for all dialog buttons

buttonSelectedAttr :: AttrName

The attribute for the selected dialog button (extends dialogAttr)

Rendering

data Result n :: * -> *

The type of result returned by a widget's rendering function. The result provides the image, cursor positions, and visibility requests that resulted from the rendering process.

emptyResult :: Result n

translateX :: Int -> Image -> Image

Translates an image by padding or cropping its left side.

translateY :: Int -> Image -> Image

Translates an image by padding or cropping its top.

charFill

Arguments

:: Integral d

=> Attr
The attribute to use.

-> Char
The character to use in filling the region.

-> d
The region width.

-> d
The region height.

-> Image

Make an image filling a region with the specified character.

If either the width or height are less than or equal to 0, then the result is the empty image.

Events & handling

data BrickEvent n e :: * -> * -> *

The type of events.

Constructors

VtyEvent Event
The event was a Vty event.

AppEvent e
The event was an application event.

MouseDown n Button [Modifier] Location
A mouse-down event on the specified region was received.

MouseUp n (Maybe Button) Location
A mouse-up event on the specified region was received.

data Event :: *

Events.

Constructors

EvKey Key [Modifier]
A keyboard key was pressed with the specified modifiers.

EvMouseDown Int Int Button [Modifier]
A mouse button was pressed at the specified column and row. Any modifiers available in the event are also provided.

EvMouseUp Int Int (Maybe Button)
A mouse button was released at the specified column and row. Some terminals report only that a button was released without specifying which one; in that case, Nothing is provided. Otherwise Just the button released is included in the event.

EvResize Int Int
If read from eventChannel this is the size at the time of the signal. If read from nextEvent this is the size at the time the event was processed by Vty. Typically these are the same, but if somebody is resizing the terminal quickly they can be different.

EvPaste ByteString
A paste event occurs when a bracketed paste input sequence is received. For terminals that support bracketed paste mode, these events will be triggered on a paste event. Terminals that do not support bracketed pastes will send the paste contents as ordinary input (which is probably bad, so beware!) Note that the data is provided in raw form and you'll have to decode (e.g. as UTF-8) if that's what your application expects.

data Key :: *

Representations of non-modifier keys.

KFun is indexed from 0 to 63. Range of supported FKeys varies by terminal and keyboard.

KUpLeft, KUpRight, KDownLeft, KDownRight, KCenter support varies by terminal and keyboard.

Actually, support for most of these but KEsc, KChar, KBS, and KEnter vary by terminal and keyboard.

Constructors

KEsc

KChar Char

KBS

KEnter

KLeft

KRight

KUp

KDown

KUpLeft

KUpRight

KDownLeft

KDownRight

KCenter

KFun Int

KBackTab

KPrtScr

KPause

KIns

KHome

KPageUp

KDel

KEnd

KPageDown

KBegin

KMenu

data Next a :: * -> *

The type of actions to take upon completion of an event handler.

continue :: s -> EventM n (Next s)

Continue running the event loop with the specified application state.

halt :: s -> EventM n (Next s)

Halt the event loop and return the specified application state as the final state value.