Exercises.Vty

Exercises

Exercise 1

plainHelloWorld :: Image

Print "Hello World" in terminal default style.

renderVtyImage plainHelloWorld

Exercise 2

colorfulHelloWorld :: Image

Print "Hello World" in red on blue

renderVtyImage colorfulHelloWorld

Exercise 3

staticVtyList :: Image

Print a list with "Item 1", "Item 2", ..., "Item 10", where the first item is in standout (reverse video) style.

renderVtyImage staticVtyList

Exercise 4

vtyListApp :: MiniApp ([String], Int)

This list is similar to list1, but should react to arrow keys by moving the index of the selected item. The internal state ([String], Int) is the list of items to be displayed, and the index of the currently selected item.

runVtyApp vtyListApp vtyListAppState

vtyListAppState :: ([String], Int)

Utilities

renderVtyImage :: Image -> IO ()

Renders the given Image full-screen, and exits on the first event:

renderVtyImage (string defAttr "Hello World")

data MiniApp s

Constructors

MiniApp

Fields

appRender :: s -> Image

How to render the image for a given state

appHandle :: Event -> s -> Maybe s

State transitions for a given event. Return Nothing to end the application.

runVtyApp :: MiniApp s -> s -> IO ()

Runs a MiniApp given an initial state

Library functions

Creating and combining Images

string :: Attr -> String -> Image

Make an Image from a String.

This is an alias for iso10646String since the usual case is that a literal string like "foo" is represented internally as a list of ISO 10646 31 bit characters.

Note: Keep in mind that GHC will compile source encoded as UTF-8 but the literal strings, while UTF-8 encoded in the source, will be transcoded to a ISO 10646 31 bit characters runtime representation.

horizCat :: [Image] -> Image

Compose any number of images together horizontally, with the first in the list being leftmost.

vertCat :: [Image] -> Image

Compose any number of images vertically, with the first in the list being topmost.

(<|>) :: Image -> Image -> Image infixr 5

Combines two images horizontally. This is an alias for horizJoin.

infixr 5

(<->) :: Image -> Image -> Image infixr 4

Combines two images vertically. This is an alias for vertJoin.

infixr 4

Attributes

defAttr :: Attr

Sets the style, background color and foreground color to the default values for the terminal. There is no easy way to determine what the default background and foreground colors are.

withForeColor :: Attr -> Color -> Attr

Set the foreground color of an Attr.

withBackColor :: Attr -> Color -> Attr

Set the background color of an Attr.

red :: Color

Standard 8-color ANSI terminal color codes.

blue :: Color

Standard 8-color ANSI terminal color codes.

withStyle :: Attr -> Style -> Attr

Add the given style attribute

standout :: Style

The 6 possible style attributes:

standout

underline

reverseVideo

blink

dim

bold/bright

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

bold :: Style

The 6 possible style attributes:

standout

underline

reverseVideo

blink

dim

bold/bright

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

Events

data Event :: *

Events.

Constructors

EvKey Key [Modifier]
A keyboard key was pressed with the specified modifiers.

EvMouseDown Int Int Button [Modifier]
A mouse button was pressed at the specified column and row. Any modifiers available in the event are also provided.

EvMouseUp Int Int (Maybe Button)
A mouse button was released at the specified column and row. Some terminals report only that a button was released without specifying which one; in that case, Nothing is provided. Otherwise Just the button released is included in the event.

EvResize Int Int
If read from eventChannel this is the size at the time of the signal. If read from nextEvent this is the size at the time the event was processed by Vty. Typically these are the same, but if somebody is resizing the terminal quickly they can be different.

EvPaste ByteString
A paste event occurs when a bracketed paste input sequence is received. For terminals that support bracketed paste mode, these events will be triggered on a paste event. Terminals that do not support bracketed pastes will send the paste contents as ordinary input (which is probably bad, so beware!) Note that the data is provided in raw form and you'll have to decode (e.g. as UTF-8) if that's what your application expects.

data Key :: *

Representations of non-modifier keys.

KFun is indexed from 0 to 63. Range of supported FKeys varies by terminal and keyboard.

KUpLeft, KUpRight, KDownLeft, KDownRight, KCenter support varies by terminal and keyboard.

Actually, support for most of these but KEsc, KChar, KBS, and KEnter vary by terminal and keyboard.

Constructors

KEsc

KChar Char

KBS

KEnter

KLeft

KRight

KUp

KDown

KUpLeft

KUpRight

KDownLeft

KDownRight

KCenter

KFun Int

KBackTab

KPrtScr

KPause

KIns

KHome

KPageUp

KDel

KEnd

KPageDown

KBegin

KMenu