Pong.Opponent

Documentation

mvOpponent :: GameState -> GameState

Simple computer opponent.

This opponent will always try to keep the paddle level with the ball, but is limited in its motion speed.