Pong.Physics

Documentation

inertialMovement :: GameState -> GameState

Move the ball inertially, i.e. along the velocity vector.

collisions :: GameState -> BoundsCheck

Check for collisions, and bounce the ball from paddles and/or bounds if necessary.

data BoundsCheck

Constructors

LeftOutOfBounds
Ball left the field on the left side (i.e. right player wins)

RightOutOfBounds
Ball left the field on the right side (i.e. left player wins)

WithinBounds GameState
Ball is inside the field (i.e. game continues)